requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
boto3==1.42.96
botocore==1.42.96
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.3
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
jmespath==1.1.0
mangum==0.21.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
s3transfer==0.16.1
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
